+++ src/main/java/org/logicng/functions/LiteralProfileFunction.java
@@ -47,10 +47,28 @@ import java.util.TreeMap;
  * For this function, the non-caching version is preferred since it usually performs better.  The non-caching version
  * of this function generates the result mapping only once and fills it recursively whereas the caching version has to
  * construct a new mapping for each sub-formula.
- * @version 2.0.0
+ * @version 2.3.0
  * @since 1.0
  */
 public final class LiteralProfileFunction implements FormulaFunction<Map<Literal, Integer>> {
+    private static final LiteralProfileFunction INSTANCE = new LiteralProfileFunction();
+
+    /**
+     * @deprecated In the next version, the standard constructor will be replaced by a private constructor.
+     * In order to instantiate an object of this class, use the {@link #get()} method.
+     */
+    @Deprecated
+    public LiteralProfileFunction() {
+        // Intentionally left empty
+    }
+
+    /**
+     * Returns the singleton instance of this function.
+     * @return an instance of this function
+     */
+    public static LiteralProfileFunction get() {
+        return INSTANCE;
+    }
 
     /**
      * The non-caching implementation of the literal profile computation.  In this case the result map is only