+++ src/main/java/org/logicng/knowledgecompilation/bdds/functions/LngBDDFunction.java
@@ -42,10 +42,29 @@ import java.util.stream.Collectors;
 
 /**
  * Creates a LogicNG internal BDD data structure of a given BDD.
- * @version 2.0.0
+ * @version 2.3.0
  * @since 2.0.0
  */
 public final class LngBDDFunction implements BDDFunction<BDDNode> {
+
+    private static final LngBDDFunction INSTANCE = new LngBDDFunction();
+
+    /**
+     * @deprecated In the next version, the standard constructor will be replaced by a private constructor.
+     * In order to instantiate an object of this class, use the {@link #get()} method.
+     */
+    @Deprecated
+    public LngBDDFunction() {
+        // intentionally left empty
+    }
+
+    /**
+     * Returns the singleton instance of this function.
+     * @return an instance of this function
+     */
+    public static LngBDDFunction get() {
+        return INSTANCE;
+    }
 
     @Override
     public BDDNode apply(final BDD bdd) {