+++ src/main/java/org/logicng/solvers/SATSolver.java
@@ -87,9 +87,9 @@ public abstract class SATSolver {
     /**
      * Adds a formula to the solver, but sets all variables to false which are not known to the solver.
      * @param formula the formula
-     * @deprecated Due to simplifications of formulas in LogicNG the solver might not know all original variables of the formulas added. If such a variable is contained in the
-     * given formula, this method sets the variable to false. For a caller this behavior is misleading and can yield to erroneous results. Instead of using this method, a caller
-     * should track the unknown/forbidden variables by itself and add each variable negated to the solver.
+     * @deprecated Due to simplifications of formulas in LogicNG the solver might not know all original variables of the formulas added. If such a variable
+     * is contained in the given formula, this method sets the variable to false. For a caller this behavior is misleading and can yield to erroneous results.
+     * Instead of using this method, a caller should track the unknown/forbidden variables by itself and add each variable negated to the solver.
      */
     @Deprecated
     public abstract void addWithoutUnknown(final Formula formula);
@@ -174,8 +174,10 @@ public abstract class SATSolver {
      * Usage constraints:
      * - "&lt;": Cannot be used with right hand side 2, returns null for right hand side 1, but constraint is added to solver.
      * - "&lt;=": Cannot be used with right hand side 1, returns null for right hand side 0, but constraint is added to solver.
-     * - "&gt;": Returns null for right hand side 0 or number of variables -1, but constraint is added to solver. Adds false to solver for right hand side &gt;= number of variables.
-     * - "&gt;=": Returns null for right hand side 1 or number of variables, but constraint is added to solver. Adds false to solver for right hand side &gt; number of variables.
+     * - "&gt;": Returns null for right hand side 0 or number of variables -1, but constraint is added to solver. Adds false to solver for right hand side
+     * &gt;= number of variables.
+     * - "&gt;=": Returns null for right hand side 1 or number of variables, but constraint is added to solver. Adds false to solver for right hand side &gt;
+     * number of variables.
      * @param cc the cardinality constraint
      * @return the incremental data of this constraint, or null if the right hand side of cc is 1
      */
@@ -415,8 +417,9 @@ public abstract class SATSolver {
     public abstract SortedSet<Variable> knownVariables();
 
     /**
-     * Returns an unsat core of the current problem.  Only works if the SAT solver is configured to record the information
+     * Returns an unsat core of the current problem. Only works if the SAT solver is configured to record the information
      * required to generate a proof trace and an unsat core.
+     * In particular, this method returns the unsat core only if the parameter "proofGeneration" in the MiniSatConfig is set to "true".
      * @return the unsat core
      */
     public UNSATCore<Proposition> unsatCore() {