+++ src/main/java/org/logicng/transformations/AIGTransformation.java
@@ -44,16 +44,34 @@ import java.util.LinkedHashSet;
 
 /**
  * And-inverter-graph (AIG) transformation.  Returns the AIG of the given formula.
- * @version 1.0
+ * @version 2.3.0
  * @since 1.0
  */
 public final class AIGTransformation implements FormulaTransformation {
 
     private FormulaFactory f;
     private boolean cache;
+    private static final AIGTransformation INSTANCE = new AIGTransformation();
+
+    /**
+     * @deprecated In the next version, the standard constructor will be replaced by a private constructor.
+     * In order to instantiate an object of this class, use the {@link #get()} method.
+     */
+    @Deprecated
+    public AIGTransformation() {
+        // Intentionally left empty
+    }
+
+    /**
+     * Returns the singleton instance of this function.
+     * @return an instance of this function
+     */
+    public static AIGTransformation get() {
+        return INSTANCE;
+    }
 
     @Override
-    public Formula apply(final Formula formula, boolean cache) {
+    public Formula apply(final Formula formula, final boolean cache) {
         this.f = formula.factory();
         this.cache = cache;
         switch (formula.type()) {
@@ -81,8 +99,8 @@ public final class AIGTransformation implements FormulaTransformation {
     private Formula transformNot(final Not not) {
         Formula aig = not.transformationCacheEntry(AIG);
         if (aig == null) {
-            aig = f.not(apply(not.operand(), cache));
-            if (cache) {
+            aig = this.f.not(apply(not.operand(), this.cache));
+            if (this.cache) {
                 not.setTransformationCacheEntry(AIG, aig);
                 aig.setPredicateCacheEntry(PredicateCacheEntry.IS_AIG, true);
             }
@@ -93,8 +111,8 @@ public final class AIGTransformation implements FormulaTransformation {
     private Formula transformImplication(final Implication impl) {
         Formula aig = impl.transformationCacheEntry(AIG);
         if (aig == null) {
-            aig = f.not(f.and(apply(impl.left(), cache), f.not(apply(impl.right(), cache))));
-            if (cache) {
+            aig = this.f.not(this.f.and(apply(impl.left(), this.cache), this.f.not(apply(impl.right(), this.cache))));
+            if (this.cache) {
                 impl.setTransformationCacheEntry(AIG, aig);
                 aig.setPredicateCacheEntry(PredicateCacheEntry.IS_AIG, true);
             }
@@ -105,9 +123,9 @@ public final class AIGTransformation implements FormulaTransformation {
     private Formula transformEquivalence(final Equivalence equiv) {
         Formula aig = equiv.transformationCacheEntry(AIG);
         if (aig == null) {
-            aig = f.and(f.not(f.and(apply(equiv.left(), cache), f.not(apply(equiv.right(), cache)))),
-                    f.not(f.and(f.not(equiv.left()), equiv.right())));
-            if (cache) {
+            aig = this.f.and(this.f.not(this.f.and(apply(equiv.left(), this.cache), this.f.not(apply(equiv.right(), this.cache)))),
+                    this.f.not(this.f.and(this.f.not(equiv.left()), equiv.right())));
+            if (this.cache) {
                 equiv.setTransformationCacheEntry(AIG, aig);
                 aig.setPredicateCacheEntry(PredicateCacheEntry.IS_AIG, true);
             }
@@ -120,10 +138,10 @@ public final class AIGTransformation implements FormulaTransformation {
         if (aig == null) {
             final LinkedHashSet<Formula> nops = new LinkedHashSet<>(and.numberOfOperands());
             for (final Formula op : and) {
-                nops.add(apply(op, cache));
+                nops.add(apply(op, this.cache));
             }
-            aig = f.and(nops);
-            if (cache) {
+            aig = this.f.and(nops);
+            if (this.cache) {
                 and.setTransformationCacheEntry(AIG, aig);
                 aig.setPredicateCacheEntry(PredicateCacheEntry.IS_AIG, true);
             }
@@ -136,10 +154,10 @@ public final class AIGTransformation implements FormulaTransformation {
         if (aig == null) {
             final LinkedHashSet<Formula> nops = new LinkedHashSet<>(or.numberOfOperands());
             for (final Formula op : or) {
-                nops.add(f.not(apply(op, cache)));
+                nops.add(this.f.not(apply(op, this.cache)));
             }
-            aig = f.not(f.and(nops));
-            if (cache) {
+            aig = this.f.not(this.f.and(nops));
+            if (this.cache) {
                 or.setTransformationCacheEntry(AIG, aig);
                 aig.setPredicateCacheEntry(PredicateCacheEntry.IS_AIG, true);
             }