+++ src/main/java/org/logicng/transformations/UnitPropagation.java
@@ -43,10 +43,29 @@ import java.util.List;
 
 /**
  * A formula transformation which performs unit propagation.
- * @version 1.3
+ * @version 2.3.0
  * @since 1.2
  */
 public final class UnitPropagation implements FormulaTransformation {
+
+    private static final UnitPropagation INSTANCE = new UnitPropagation();
+
+    /**
+     * @deprecated In the next version, the standard constructor will be replaced by a private constructor.
+     * In order to instantiate an object of this class, use the {@link #get()} method.
+     */
+    @Deprecated
+    public UnitPropagation() {
+        // Intentionally left empty
+    }
+
+    /**
+     * Returns the singleton instance of this function.
+     * @return an instance of this function
+     */
+    public static UnitPropagation get() {
+        return INSTANCE;
+    }
 
     @Override
     public Formula apply(final Formula formula, final boolean cache) {