+++ src/main/java/org/logicng/transformations/simplification/BackboneSimplifier.java
@@ -42,10 +42,29 @@ import org.logicng.solvers.functions.BackboneFunction;
  * it through the formula. E.g. in the formula {@code A & B & (A | B | C) & (~B | D)}
  * the backbone {@code A, B} is computed and propagated, yielding the simplified formula
  * {@code A & B & D}.
- * @version 1.5.0
+ * @version 2.3.0
  * @since 1.5.0
  */
 public final class BackboneSimplifier implements FormulaTransformation {
+
+    private static final BackboneSimplifier INSTANCE = new BackboneSimplifier();
+
+    /**
+     * @deprecated In the next version, the standard constructor will be replaced by a private constructor.
+     * In order to instantiate an object of this class, use the {@link #get()} method.
+     */
+    @Deprecated
+    public BackboneSimplifier() {
+        // Intentionally left empty
+    }
+
+    /**
+     * Returns the singleton instance of this function.
+     * @return an instance of this function
+     */
+    public static BackboneSimplifier get() {
+        return INSTANCE;
+    }
 
     @Override
     public Formula apply(final Formula formula, final boolean cache) {